% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charXYnumXY.R
\name{charXYnumXY}
\alias{charXYnumXY}
\title{Convert Character x,y Data to Numeric x, y Values}
\usage{
charXYnumXY(VL)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector.
       Depending upon the \code{mode} and the stage of processing:
\itemize{
  \item the lines may be named as "Line_1" etc
  \item a format code may be present as the first entry
  \item there may be other non-numeric lines that eventually get stripped off
}}
}
\value{
A data frame with elements \code{x} and \code{y}.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
