% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompressXYY.R
\name{decompressXYY}
\alias{decompressXYY}
\title{Extract the values in JCAMP-DX file with an XYY variable list.}
\usage{
decompressXYY(VL, params, mode, SOFC, debug = 0)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector. Includes one pre-pended
line giving the fmt of data (e.g. XYY, XRR, XII).}

\item{params}{Numeric. Vector of parameters extracted from file header.}

\item{mode}{Character. One of c("IR_etc", "NMR", "NMR2D")}

\item{SOFC}{Logical.  See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A data frame with elements \code{x} and \code{y}.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
