% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deDIF.R
\name{deDIF}
\alias{deDIF}
\title{Process DIF Encoded Strings}
\usage{
deDIF(i, lineList, debug = 0)
}
\arguments{
\item{i}{Integer. Essentially the index into \code{lineList}.}

\item{lineList}{A list of numeric vectors to be checked.  Named with line numbers.
Individual numbers are named with the ASDF code.  X values still present!}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A numeric vector.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
