% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParams.R
\name{extractParams}
\alias{extractParams}
\title{Extract Parameters from Metadata}
\usage{
extractParams(md, mode, SOFC, debug = 0)
}
\arguments{
\item{md}{Character.  A vector of character strings which contains the metadata.}

\item{mode}{Character. One of c("IR", "NMR", "NMR2D").}

\item{SOFC}{Logical. Stop on Failed Check.   See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A named numeric vector containing the extracted parameters.
Contents will vary by \code{mode}.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
