% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariableLists.R
\name{findVariableLists}
\alias{findVariableLists}
\title{Locate Variable Lists and Related Information in a JCAMP-DX File}
\usage{
findVariableLists(jdx, debug = 0)
}
\arguments{
\item{jdx}{Character.  A vector of character strings which hopefully contains
one or more variable lists.  Each string is a line of the complete original file.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A list.
\itemize{
  \item A data frame giving information about the structure of the file. Columns will be
        Format, FirstLine, LastLine.  Serves as a Data Guide.
  \item The metadata
  \item The line numbers of comments (excluding comments in the metadata).
  \item Each variable list that was found, with the format pre-pended.
 }
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
