% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComp.R
\name{getComp}
\alias{getComp}
\title{Determine the Type(s) of Compression in Use}
\usage{
getComp(VL, debug = 0)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector.
       Depending upon the \code{mode} and the stage of processing:
\itemize{
  \item the lines may be named as "Line_1" etc
  \item a format code may be present as the first entry
  \item there may be other non-numeric lines that eventually get stripped off
}}

\item{debug}{Integer. See \code{\link{readJDX}} for details.}
}
\value{
A string giving all compression formats found.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
