% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJDXcompression.R
\name{getJDXcompression}
\alias{getJDXcompression}
\title{Determine the type(s) of Compression in Use}
\usage{
getJDXcompression(string, debug = 0)
}
\arguments{
\item{string}{Character.  The string to be checked.  If a vector of
strings it will be collapsed to a single string.}

\item{debug}{Integer. See \code{\link{readJDX}} for details.}
}
\value{
A string giving all compression formats found.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
