% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDUPs.R
\name{insertDUPs}
\alias{insertDUPs}
\title{Insert Duplicate Entries}
\usage{
insertDUPs(line, debug)
}
\arguments{
\item{line}{A character vector. Entries are named with the ASDF compression mode.  X values still present.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A string representing a processed line.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
