% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process2DNMR.R
\name{process2DNMR}
\alias{process2DNMR}
\title{Extract the x, y Values in a 2D NMR Variable List}
\usage{
process2DNMR(VL, params, mode, SOFC, debug = 0)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector.
       Depending upon the \code{mode} and the stage of processing:
\itemize{
  \item the lines may be named as "Line_1" etc
  \item a format code may be present as the first entry
  \item there may be other non-numeric lines that eventually get stripped off
}}

\item{params}{Numeric. Vector of parameters extracted from file header.}

\item{mode}{Character. One of c("IR_etc", "NMR", "NMR2D").  Not the same as "fmt" (see notes in code).
Determines the needed processing.}

\item{SOFC}{Logical. Stop on Failed Check.   See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A data frame with elements \code{x} and \code{y}.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
\section{Note}{

This function is analogous to \code{\link{processLCMS}} but differs in detail.
}

