% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processLCMS.R
\name{processLCMS}
\alias{processLCMS}
\title{Extract the x, y Values in a LC-MS Variable List}
\usage{
processLCMS(VL, SOFC, debug = 0)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector.
       Depending upon the \code{mode} and the stage of processing:
\itemize{
  \item the lines may be named as "Line_1" etc
  \item a format code may be present as the first entry
  \item there may be other non-numeric lines that eventually get stripped off
}}

\item{SOFC}{Logical. Stop on Failed Check.   See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A data frame with elements \code{mz} and \code{int}.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
\section{Note}{

This function is analogous to \code{\link{process2DNMR}} but differs in detail.
}

