% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariableLIst.R
\name{processVariableList}
\alias{processVariableList}
\title{Process a Single Variable List from a JCAMP-DX file}
\usage{
processVariableList(VL, params, mode, lineNos, SOFC, debug = 0)
}
\arguments{
\item{VL}{Character.  The variable list to be processed as a character vector.
       Depending upon the \code{mode} and the stage of processing:
\itemize{
  \item the lines may be named as "Line_1" etc
  \item a format code may be present as the first entry
  \item there may be other non-numeric lines that eventually get stripped off
}}

\item{params}{Numeric. Vector of parameters extracted from file header.}

\item{mode}{Character. One of c("IR_etc", "NMR", "NMR2D").  Not the same as "fmt" (see notes in code).
Determines the needed processing.}

\item{lineNos}{Two integers giving the first and last lines
of the variable list in the original file. Used to label debugging responses.}

\item{SOFC}{Logical. Stop on Failed Check.   See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A data frame with elements \code{x} and \code{y}, unless the file contains 2D NMR, in which case a matrix.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
