% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDUPs.R
\name{repDUPs}
\alias{repDUPs}
\title{Replicate Duplicate Values}
\usage{
repDUPs(string, dupstr)
}
\arguments{
\item{string}{Character.  A string to be replicated, as type character.
May contain a number or a character or a combination, but the whole is a
character string.}

\item{dupstr}{Character.  A character string which when decoded indicates
the number of times to repeat \code{string}.  Names are added by the calling function.}
}
\value{
An expanded string.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
