% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unDIF.R
\name{unDIF}
\alias{unDIF}
\title{Convert from DIF Format to a Numeric Vector}
\usage{
unDIF(string)
}
\arguments{
\item{string}{Character. A character vector to be converted. Named with compression codes.}
}
\value{
A numeric vector with the final values, named with the compression codes.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
