% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unSQZ.R
\name{unSQZ}
\alias{unSQZ}
\title{Translate SQZ Codes}
\usage{
unSQZ(string)
}
\arguments{
\item{string}{A character string containing a SQZ code}
}
\value{
A charcter string in which the SQZ letter code has been replaced by a number.
}
\description{
Translate SQZ Codes
}
