% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yValueCheck.R
\name{yValueCheck}
\alias{yValueCheck}
\title{Conditionally Carry Out the y Value Check}
\usage{
yValueCheck(lineList, debug = 0)
}
\arguments{
\item{lineList}{A list of numeric vectors to be checked.  Named with line numbers.
Individual numbers are named with the ASDF code.  X values are still present!}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A list of numeric vectors, after the y check has been done and any extra
        values removed. It will stop if there is a problem.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
