% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompLines.R
\name{decompLines}
\alias{decompLines}
\title{Decompress the Data Lines in a JCAMP-DX File}
\usage{
decompLines(lines, debug = 0)
}
\arguments{
\item{lines}{Character.  A vector of character strings composed of the compressed
data strings in the file.  If coming from the package internals, named as "Line_1" etc.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A list of numeric strings, the result of unpacking the compressed data.  One list
        element is returned for each line.  The numeric string is named with the compression mode.
}
\description{
This function is used by the package, and can also be called by the user who wishes
to study problematic lines in a file.
}
\section{Details}{

         Each individual line passed here is converted to a list to make processing and
         naming the pieces easier.  \emph{No checking of values is done here}.  The individual
         numbers are named according to their \emph{original} compression code.  Unless otherwise noted
         all functions called from here should act on a list of lines, via lapply. Note too that the x
         values are always present here, so X, Y1, Y2 ...
}

\section{Formats}{

         AFFN is separated by any amount of white space so processing is straightforward, as exponents are
         handled automatically, and white space stripped off automatically, courtesy of R internals.  It appears
         AFFN is never mixed with other formats.  The other formats are collectively called ASDF in the standard.
}

