% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDUPs.R
\name{insertDUPs}
\alias{insertDUPs}
\title{Insert Duplicate Entries}
\usage{
insertDUPs(string, debug)
}
\arguments{
\item{string}{Character vector.  Partially processed lines from the
original variable list.  Should arrive here named by compression code.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A string.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
