% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processVariableLIst.R
\name{processVariableList}
\alias{processVariableList}
\title{Decompress a Single Variable List from a JCAMP-DX file}
\usage{
processVariableList(VL, params, mode, lineNos, SOFC, debug = 0)
}
\arguments{
\item{VL}{Character.  A vector of character strings which contains
the variable list.  First line should be the format code (an
extra line inserted by this package).  Depending upon mode, there
may be some other stuff that still needs to be stripped off to get to just numbers.}

\item{params}{Numeric. Vector of parameters from file header.}

\item{mode}{Character. One of c("IR_etc", "NMR", "NMR2D")}

\item{lineNos}{Two integers giving the first and last lines
of the variable list in the original file. Used for debugging responses.}

\item{SOFC}{Logical.  See \code{\link{readJDX}} for details.}

\item{debug}{Integer.  See \code{\link{readJDX}} for details.}
}
\value{
A data frame with elements $x$ and $y$, unless the input is 2D NMR, in which case a matrix.
}
\description{
This function is NOT EXPORTED.
Users would not normally call this function.  See \code{\link{readJDX}}.
Documentation is provided for developers wishing to contribute to the package.
}
