% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMultiblockDX.R
\name{splitMultiblockDX}
\alias{splitMultiblockDX}
\title{Split a Multiblock JCAMP-DX File into Individual Files}
\usage{
splitMultiblockDX(file = NULL)
}
\arguments{
\item{file}{Length one character giving the name of the multiblock file.}
}
\description{
Process a multiblock JCAMP-DX file into separate files, one for each data set.
}
\section{Warning}{

\pkg{Individual data sets are written into the current directory, so you probably want to create
a dedicated folder for this task}.
}

\author{
Bryan A. Hanson, DePauw University.
}
\keyword{file}
\keyword{utilities}
